% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianPolishM.ConstructMPst.R
\name{MedianPolishM.ConstructMPst}
\alias{MedianPolishM.ConstructMPst}
\title{Median polish multidimensional.}
\usage{
\method{MedianPolishM}{ConstructMPst}(data, eps, maxiter, na.rm, ...)
}
\arguments{
\item{data}{class \code{\link{ConstructMPst}}.}

\item{eps}{real number greater than \code{0}, default 0.01. A tolerance for convergence: see Details}

\item{maxiter}{the maximum number of iterations. Default 10.}

\item{na.rm}{logical. If the data contains NA's. Default TRUE.}

\item{\dots}{ignored.}
}
\value{
An object of class medpolish with the following named components in a list:

\item{residuals}{the residuals.}

\item{overall}{the fitted constant term.}

\item{effects}{the fitted every space - time effects.}

\item{iter}{number of iterations used in the range maxiter.}
}
\description{
An additive model for multidimensional array is fitted, using Tukey's median polish procedure.
}
\details{
The model fitted is an additive, \eqn{\mu +\alpha _{a} + \beta _{b} + \xi _{c} + \tau _{t}}, where \eqn{\mu} is an overall mean, \eqn{\alpha_{a}} is the \eqn{a}-th row effect, \eqn{\beta_{b}} is the effect \eqn{b}-th column effect, \eqn{\xi_{c}} is the \eqn{c}-th layer effect, \eqn{\tau _{t}} is the \eqn{t}-th time effect. The algorithm works by alternately removing medians of every spatio - temporal dimensions, and continues until the proportional reduction in the sum of absolute residuals is less than eps or until there have been maxiter iterations. If na.rm is FALSE, the presence of any NA value in x will cause an error, otherwise NA values are ignored. MedianPolishM returns an object of class MedianPolishM (see below). There is a plotting method for this class, \code{\link{plot.MedianPolishM}}.
}
\references{
Martínez, W. A., Melo, C. E., & Melo, O. O. (2017). \emph{Median Polish Kriging for space--time analysis of precipitation} Spatial Statistics, 19, 1-20. \href{http://www.sciencedirect.com/science/article/pii/S2211675316301336}{[link]}

Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (Eds.). (2011). \emph{Exploring data tables, trends, and shapes} (Vol. 101). John Wiley & Sons.\href{http://www.wiley.com/WileyCDA/WileyTitle/productCd-047004005X.html}{[link]}
}
