% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MedianPolishM.R
\name{MedianPolishM}
\alias{MedianPolishM}
\title{Median polish multidimensional.}
\usage{
MedianPolishM(data, ...)
}
\arguments{
\item{data}{object of class array, table, or matrix, see details.}

\item{\dots}{default arguments, see \code{\link{MedianPolishM.default}}}
}
\value{
An object of class medpolish with the following named components in a list:

\item{residuals}{the residuals.}

\item{overall}{the fitted constant term.}

\item{effects}{the fitted every dimensions effects fo array multidimensional.}

\item{iter}{number of iterations used in the range maxiter.}
}
\description{
Fits an additive model for multidimensional array, using Tukey's median polish procedure.
}
\details{
The function MedianPolishM is generic. See the documentation for \code{\link{MedianPolishM.default}} for further details.
}
\references{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (Eds.). (2011). \emph{Exploring data tables, trends, and shapes} (Vol. 101). John Wiley & Sons.\href{http://www.wiley.com/WileyCDA/WileyTitle/productCd-047004005X.html}{[link]}
}

