\name{Metadb}
\alias{Metadb}
\docType{data}
\title{
Monthly precipitation Meta.
}
\description{
Records of 102 pluviometrics station of the 'Instituto de Hidrologia, Meteorologia y Estudios Ambientales de Colombia' (IDEAM), to the west of hidrological zone Meta river. Every station has 37 records of the precipitation monthly from january 2007 to january 2010.
}
\usage{data(Metadb)}
\format{
  The format is:
formal class 'data.frame'
}
\source{
\url{http://www.ideam.gov.co/}
}
\examples{
data(Metadb)
str(Metadb)
names(Metadb)
}
\keyword{datasets}
