% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.MedianPolishM.R
\name{plot.MedianPolishM}
\alias{plot.MedianPolishM}
\title{Plot Median polish multidimensional.}
\usage{
\method{plot}{MedianPolishM}(x, ...)
}
\arguments{
\item{x}{object of class MedianPolishM.}

\item{\dots}{ignored.}
}
\description{
Plot the effects of an additive model for multidimensional array, using Tukey's median polish procedure.
}
\details{
The object of class MedianPolish has a list of components of effects that allow to graphic after each iterations, the behavior of this components. If the medianpolish is apply to data of class ConstructutMPst, this method has a specific graphic for data with space - time variability.
}
\examples{
A<-MedianPolishM(UCBAdmissions, eps=0.1, maxiter=2, na.rm=TRUE)
plot(A)
}
\references{
Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (Eds.). (2011). \emph{Exploring data tables, trends, and shapes} (Vol. 101). John Wiley & Sons.\href{http://www.wiley.com/WileyCDA/WileyTitle/productCd-047004005X.html}{[link]}
}

