% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianPolishM.R
\name{MedianPolishM}
\alias{MedianPolishM}
\title{Median polish multidimensional.}
\usage{
MedianPolishM(data, ...)
}
\arguments{
\item{data}{object of class array, table or matrix (see details).}

\item{\dots}{default arguments, see \code{\link{MedianPolishM.default}}}
}
\value{
An object of class medpolish with the following named components in a list:

\item{residuals}{the residuals.}

\item{overall}{the fitted constant term.}

\item{effects}{the fitted every dimensions effects to array multidimensional.}

\item{iter}{number of iterations used in the range maxiter.}
}
\description{
An additive model for multidimensional array is fitted, using Tukey's median polish procedure.
}
\details{
The function MedianPolishM is generic. See the documentation for \code{\link{MedianPolishM.default}} for further details.
}
\references{
Martínez, W. A., Melo, C. E., & Melo, O. O. (2017). \emph{Median Polish Kriging for space--time analysis of precipitation} Spatial Statistics, 19, 1-20. \href{http://www.sciencedirect.com/science/article/pii/S2211675316301336}{[link]}

Hoaglin, D. C., Mosteller, F., & Tukey, J. W. (Eds.). (2011). \emph{Exploring data tables, trends, and shapes} (Vol. 101). John Wiley & Sons.\href{http://www.wiley.com/WileyCDA/WileyTitle/productCd-047004005X.html}{[link]}
}
